models/NaiveBayes/requirements.txt
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.0
mlflow==2.12.1
spacy==3.7.2
python-dotenv==1.0.0
fsspec
huggingface_hub
boto3